Node: add {get,find}ChildNamed() methods
Both methods look for a child with the given name as an immediate child of
the current node.

findChildNamed:

Will return an Optional containing the found node or empty.

getChildNames:

Is more insistent and will return the found node itself. If a node by that
name is not found, then a NodeException will be thrown.